Move and re-export macro
